fix: adopt Copilot suggestions for type safety and buffer count guard
- Use explicit Int64 conversion in lastOffset comparison for clarity
- Require 3 buffers (not 2) for variable-width types since downstream
  code accesses buffers[2] for the values buffer

Made-with: Cursor

+++ Sources/Arrow/ArrowReader.swift
@@ -202,7 +202,7 @@ public class ArrowReader { // swiftlint:disable:this type_body_length
         guard lastOffset >= 0 else {
             return .failure(.invalid("Negative last offset (\(lastOffset)) in variable-width buffer"))
         }
-        guard lastOffset <= valueBuffer.length else {
+        guard Int64(lastOffset) <= valueBuffer.length else {
             return .failure(.invalid(
                                 "Last offset (\(lastOffset)) exceeds value buffer length (\(valueBuffer.length))"))
         }
